Move quest editing into header component
